Redis Cache: move RedisCacheImplTest to the deployment module
